Project-Files/project/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
python-multipart==0.0.6
ibm-watsonx-ai==1.0.2
PyMuPDF==1.23.8
langchain==0.0.335
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
plotly==5.17.0
streamlit-chat==0.1.1
streamlit-option-menu==0.3.6